Removed old unused pages
