+++ client/src/js/BaseComponent.js
@@ -1,0 +1,8 @@
+module.exports = class BaseComponent {
+  constructor(props) {
+    this.props = props;
+  }
+  template() {
+    throw new Error('Unimplemented template');
+  }
+}
