--- client/src/js/config/config.js
@@ -1,5 +1,0 @@
-
-
-const Config = {};
-
-module.exports = Config;
