--- client/src/js/config/watchNotificationsFormConfig.js
@@ -1,20 +1,0 @@
-
-
-const notificationsConfig = [
-  {
-    label: 'Go to contacts on right button',
-    notificationType: 'dummyNotification',
-    defaultValue: 'Click right button to go to contacts.',
-  },
-  {
-    label: 'Left button will be contacts now',
-    notificationType: 'dummyNotification',
-    defaultValue: 'Now left button will lead you to contacts.',
-  },
-  {
-    label: 'No default message!',
-    notificationType: 'dummyNotification',
-  },
-];
-
-module.exports = notificationsConfig;
