+++ client/src/js/main.js
@@ -4,12 +4,9 @@ require('../fonts/fonts.scss');
 const routes = require('./routes');
 const App = require('./app');
 const NotificationsForm = require('watch_framework').NotificationsForm;
-const notificationsFormConfig = require('./config/watchNotificationsFormConfig');
-
-const $ = require('jquery');
 
 const notificationsForm = new NotificationsForm();
-notificationsForm.configureNotifications(notificationsFormConfig);
+notificationsForm.configureNotifications({});
 
 const watch = {
   watchFace: document.getElementById('watch-face'),
