+++ client/src/js/routes.js
@@ -1,5 +1,5 @@
 const HomePage = require('./pages/HomePage');
-const ContactsPage = require('./pages/ContactsPage');
+const ContactsPage = require('./pages/ContactsPage/');
 const TeamPage = require('./pages/TeamPage');
 const FourOhFour = require('./pages/404Page');
