--- client/src/js/watch-notifications.js
@@ -1,6 +1,0 @@
-const watchFramework = require('watch_framework');
-
-// TODO autogenerate this file
-module.exports = {
-  dummyNotification: watchFramework.Notification,
-};
