--- client/src/templates/pages/eventDetails.hbs
@@ -1,2 +1,0 @@
-<h1>{{name}}</h1>
-<p>{{description}}</p>
